refactor: 쿠폰 목록 화면 관련 파일명 및 패키지 구조 변경
