+++ app/src/main/java/com/conkeep/ui/feature/coupon/list/component/CouponCard.kt
@@ -1,4 +1,4 @@
-package com.conkeep.ui.feature.coupon.component
+package com.conkeep.ui.feature.coupon.list.component
 
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Column
